saved variant tweaks
